casing, reduce some computation

+++ ProjectEuler/ProjectEuler/Problem6.cpp
@@ -5,6 +5,6 @@ int ProblemSix()
     const int N = 100;
     int i = 0;
     int sumOfSquares = N * (N + 1) * (2 * N + 1) / 6;
-    int squareOfSum = pow((N * (N + 1) / 2), 2);
+    int squareOfSum = (int)(pow((N * (N + 1) / 2), 2));
     return (squareOfSum - sumOfSquares);
 }
